==> code_age_no_impact_weights/scripts/01_simulate_pop/02_generate_pop_samp_design.R <==

### SET UP ----------------------------------------------------------------------

# load packages
library(tidyverse)

# clear working environment
rm(list = ls())

# create expit function to calculate probabilities 
expit <- function(x) {exp(x)/(1 + exp(x))}

### GENERATE POPULATION ---------------------------------------------------------

# set seed 
set.seed(13)

# set number of BGs
n_bg <- 752
# set number of stratas
n_strata <- 8
# set number of block groups (BGs) for each strata
n_bg_per_strata <- c(58, 58, 21, 21, 130, 130, 167, 167)
# set BG sampling probability for each strata
bg_samp_prob_per_strata <- c(rep(0.25, 4), rep(0.60, 4))
 
# create BG data frame of population where each row is a BG
df_pop_bg <- data.frame(
  # create one row for each BG
  bg = seq_len(n_bg),
  # assign correct strata for each BG
  strata = rep(seq_len(n_strata), n_bg_per_strata),  
  # assign number of households for each BG
  bg_size = 1 + round(rexp(n = 752, rate = 1/450)),
  # assign strata-specific sampling probability for each BG
  bg_sampling_prob = rep(bg_samp_prob_per_strata, n_bg_per_strata)
)

# expand BG data frame to create household (HH) data frame of population where each row is an HH and generate HH mean age
df_pop_hh <- df_pop_bg %>% 
  # create one row for each HH
  uncount(bg_size, .remove = FALSE) %>%
  # add columns for HH ID, HH size and HH mean age
  # where HH mean age is generated by sampling from N(40, 225) distribution truncated to range 23 to 69
  mutate(
    hh = seq_len(nrow(.)),
    hh_size = 2, 
    hh_mean_age = rnorm(nrow(.), mean = 40, sd = 15),
    hh_mean_age = case_when(
      hh_mean_age < 23 ~ 23,
      hh_mean_age > 69 ~ 69,
      TRUE ~ hh_mean_age
    )
  )

# stratify households by hispanic surname/no hispanic surname
df_pop_hh <- df_pop_hh %>%
  group_by(strata) %>%
  # sample different percentages of HH based on strata and assign those as Hispanic surname
  sample_frac(size = case_when(
    strata == 1 | strata == 2 | strata == 3 | strata == 4 ~ 0.25,
    strata == 5 | strata == 6 | strata == 7 | strata == 8 ~ 0.35
  )) %>%
  mutate(hh_hisp_surname = 1) %>%
  ungroup() %>%
  select(hh, hh_hisp_surname) %>% 
  # rejoin the HHs sampled as Hispanic surname to the entire HH data frame
  right_join(df_pop_hh, by = "hh") %>% 
  mutate(hh_hisp_surname = if_else(is.na(hh_hisp_surname), 0, 1)) %>%
  arrange(hh) %>%
  select("bg", "strata", "bg_size", "bg_sampling_prob", "hh", "hh_size", "hh_mean_age", "hh_hisp_surname")

# check to make sure stratification by hispanic surname was done properly
#df_pop_hh %>% 
  #group_by(strata, hh_hisp_surname) %>%
  #count() %>%
  #group_by(strata) %>%
  #mutate(total = sum(n)) %>%
  #ungroup() %>%
  #mutate(frac = n/total)

# add HH sampling probabilities stratified by hispanic surname status
df_pop_hh <- df_pop_hh %>%
  mutate(
    hh_sampling_prob = case_when(
      (strata == 1 | strata == 2) & hh_hisp_surname == 1 ~ 0.10,
      (strata == 1 | strata == 2) & hh_hisp_surname == 0 ~ 0.05,
      (strata == 3 | strata == 4) & hh_hisp_surname == 1 ~ 0.14,
      (strata == 3 | strata == 4) & hh_hisp_surname == 0 ~ 0.04,
      (strata == 5 | strata == 6) & hh_hisp_surname == 1 ~ 0.09,
      (strata == 5 | strata == 6) & hh_hisp_surname == 0 ~ 0.03,
      (strata == 7 | strata == 8) & hh_hisp_surname == 1 ~ 0.08,
      (strata == 7 | strata == 8) & hh_hisp_surname == 0 ~ 0.02
    )
  )

# check to make sure HH sampling probabilities were added correctly
#df_pop_hh %>%
  #group_by(strata, hh_hisp_surname) %>%
  #summarise(mean = mean(hh_sampling_prob))

# expand HH data frame to create individual data frame of population where each row is an individual
# and generate age for each individual and HH sampling probability and calculate max age in HH
df_pop_indiv <- df_pop_hh %>%
  # create one row for each individual
  uncount(hh_size, .remove = FALSE) %>%
  # add individual ID column
  mutate(id = seq_len(nrow(.))) %>%
  # generate individual's age by sampling from uniform, discrete distribution that ranges within 10 years of HH mean age
  # this sampling method will create a HH cluster-like effect for age
  rowwise(.) %>%
  mutate(age = sample(
    seq(round(hh_mean_age - 5), round(hh_mean_age + 5)), 
    size = 1)
  ) %>%
  ungroup(.)
